Move some similar functions from distutils.sysconfig -> sysconfig
